backend/requirements.txt
fastapi
uvicorn
numpy
yt-dlp
whisper
pydantic
pymongo
langchain_huggingface
python-dotenv
